Add files via upload
This project showcases my proficiency in front-end web development, specifically in HTML, CSS, JavaScript, and Bootstrap. It demonstrates my ability to create user-friendly and visually appealing interfaces tailored to the needs of specific user groups. I undertook the development of a student dashboard website using HTML, CSS, JavaScript, and Bootstrap. This project aimed to create a user-friendly and informative platform for students to access their academic information and resources.

Using HTML, I structured the website's content, including sections for the header, navigation menu, dashboard widgets, course listings, announcements, and student profile information. The layout was designed to be intuitive and visually appealing, ensuring ease of use for students.

CSS played a crucial role in styling the website. I leveraged CSS to customize the colors, typography, and overall visual design of the dashboard. The goal was to create a clean and modern interface that aligns with the needs and preferences of students.

To enhance interactivity and functionality, JavaScript was employed. I implemented dynamic features such as interactive widgets, collapsible panels, and form validation to provide an engaging user experience. JavaScript also facilitated real-time updates of data, such as grades and course schedules, ensuring students had access to the most up-to-date information.

Bootstrap, a widely used front-end framework, was instrumental in building a responsive and mobile-friendly website. By utilizing Bootstrap's components and responsive grid system, I ensured that the dashboard adapted seamlessly to different screen sizes and devices. This responsiveness allowed students to access their information conveniently on desktop computers, tablets, or mobile phones.